Relocate LayoutBody
