+++ client/a8c-for-agencies/components/page-placeholder/index.tsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx';
-import LayoutBody from 'calypso/a8c-for-agencies/components/layout/body';
 import Layout from 'calypso/layout/multi-sites-dashboard';
+import LayoutBody from 'calypso/layout/multi-sites-dashboard/body';
 import LayoutHeader, {
 	LayoutHeaderTitle as Title,
 } from 'calypso/layout/multi-sites-dashboard/header';
