+++ client/a8c-for-agencies/sections/overview/overview.tsx
@@ -1,9 +1,9 @@
 import { useTranslate } from 'i18n-calypso';
 import ContentSidebar from 'calypso/a8c-for-agencies/components/content-sidebar';
-import LayoutBody from 'calypso/a8c-for-agencies/components/layout/body';
 import PressableUsageLimitNotice from 'calypso/a8c-for-agencies/components/pressable-usage-limit-notice';
 import MobileSidebarNavigation from 'calypso/a8c-for-agencies/components/sidebar/mobile-sidebar-navigation';
 import Layout from 'calypso/layout/multi-sites-dashboard';
+import LayoutBody from 'calypso/layout/multi-sites-dashboard/body';
 import LayoutHeader, {
 	LayoutHeaderActions as Actions,
 	LayoutHeaderTitle as Title,
