+++ client/a8c-for-agencies/sections/purchases/payment-methods/payment-method-add/index.tsx
@@ -1,7 +1,6 @@
 import { Card } from '@automattic/components';
 import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
-import LayoutBody from 'calypso/a8c-for-agencies/components/layout/body';
 import LayoutStepper from 'calypso/a8c-for-agencies/components/layout/stepper';
 import MobileSidebarNavigation from 'calypso/a8c-for-agencies/components/sidebar/mobile-sidebar-navigation';
 import {
@@ -11,6 +10,7 @@ import {
 import PaymentMethodStripeInfo from 'calypso/jetpack-cloud/sections/partner-portal/payment-method-stripe-info';
 import { usePaymentMethodStepper } from 'calypso/jetpack-cloud/sections/partner-portal/primary/payment-method-add-v2/hooks/use-payment-method-stepper';
 import Layout from 'calypso/layout/multi-sites-dashboard';
+import LayoutBody from 'calypso/layout/multi-sites-dashboard/body';
 import LayoutHeader, {
 	LayoutHeaderTitle as Title,
 	LayoutHeaderSubtitle as Subtitle,
