+++ client/a8c-for-agencies/sections/team/primary/team-accept-invite/index.tsx
@@ -1,7 +1,6 @@
 import page from '@automattic/calypso-router';
 import { useTranslate } from 'i18n-calypso';
 import { useEffect, useMemo, useState } from 'react';
-import LayoutBody from 'calypso/a8c-for-agencies/components/layout/body';
 import PagePlaceholder from 'calypso/a8c-for-agencies/components/page-placeholder';
 import { A4A_OVERVIEW_LINK } from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
 import useActivateMemberMutation, {
@@ -9,6 +8,7 @@ import useActivateMemberMutation, {
 } from 'calypso/a8c-for-agencies/data/team/use-activate-member';
 import AgencyLogo from 'calypso/assets/images/a8c-for-agencies/agency-logo.svg';
 import Layout from 'calypso/layout/multi-sites-dashboard';
+import LayoutBody from 'calypso/layout/multi-sites-dashboard/body';
 import LayoutHeader, {
 	LayoutHeaderTitle as Title,
 } from 'calypso/layout/multi-sites-dashboard/header';
