+++ client/a8c-for-agencies/sections/team/primary/team-invite/index.tsx
@@ -9,10 +9,10 @@ import useShowFeedback from 'calypso/a8c-for-agencies/components/a4a-feedback/ho
 import Form from 'calypso/a8c-for-agencies/components/form';
 import FormField from 'calypso/a8c-for-agencies/components/form/field';
 import FormSection from 'calypso/a8c-for-agencies/components/form/section';
-import LayoutBody from 'calypso/a8c-for-agencies/components/layout/body';
 import { A4A_TEAM_LINK } from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
 import useSendTeamMemberInvite from 'calypso/a8c-for-agencies/data/team/use-send-team-member-invite';
 import Layout from 'calypso/layout/multi-sites-dashboard';
+import LayoutBody from 'calypso/layout/multi-sites-dashboard/body';
 import LayoutHeader, {
 	LayoutHeaderBreadcrumb as Breadcrumb,
 } from 'calypso/layout/multi-sites-dashboard/header';
