+++ client/layout/multi-sites-dashboard/style.scss
@@ -49,16 +49,17 @@ html, body {
 	flex: 1;
 }
 
-.multi-sites-dashboard-layout__top-wrapper {
+.a4a-layout__body {
+	width: 100%;
+	overflow-y: auto;
+	padding-block-start: 16px;
+	padding-block-end: 32px;
+}
+
+.multi-sites-dashboard-layout__top-wrapper,
+.a4a-layout__body {
 	margin-inline: 0;
 	max-height: 100%;
-	padding-block-start: 24px;
-	border-bottom: #f1f1f1 1px solid;
-
-	@include breakpoint-deprecated( "<660px" ) {
-		padding: 0;
-		border-block-end: 1px solid var(--color-neutral-5);
-	}
 
 	> * {
 		padding-inline: 16px;
@@ -68,6 +69,16 @@ html, body {
 			margin-inline: auto !important;
 			padding-inline: 64px;
 		}
+	}
+}
+
+.multi-sites-dashboard-layout__top-wrapper {
+	padding-block-start: 24px;
+	border-bottom: #f1f1f1 1px solid;
+
+	@include breakpoint-deprecated( "<660px" ) {
+		padding: 0;
+		border-block-end: 1px solid var(--color-neutral-5);
 	}
 
 	.notice-banner {
